Upgraded packages.  Changed scripts section to run in Windows environments instead of unix based environments.
