+++ tsconfig.json
@@ -2,18 +2,29 @@
     "compilerOptions": {
         "module": "commonjs",
         "moduleResolution": "node",
+        "esModuleInterop": true,
         "target": "esnext",
         "noImplicitAny": true,
         "removeComments": true,
         "allowJs": false,
         "preserveConstEnums": true,
-        "outDir": "lib/",
+        "outDir": "dist",
         "declaration": true,
-        "sourceMap": true
+        "sourceMap": true,
+        
+        
+        "baseUrl": ".",
+        "paths": {
+            "*": [
+                "node_modules/*",
+                "src/types/*"
+            ]
+        },
+        "noEmitOnError": true
     },
     "include": [
-        "src/IS.ts"
+        "src/**/*",
+        "test/**/*"
     ],
-    "exclude": [ "bin", "node_modules", "lib" ],
-    "atom": { "rewriteTsconfig": false }
+    "exclude": [ "bin", "node_modules", "dist" ]
 }
